// src/app/models/people.ts
export interface People {
  gender: string;
  hair_color: string;
  height: string;
  homeworld: string;
  mass: string;
  name: string;
  skin_color: string;
  eye_color: string;
  created: string;
  edited: string;
  species: string;
  starships: string;
  films: string;
  url: string;
  vehicles: string;
}